TP2/requirements.txt
aiohttp
beautifulsoup4
lxml
Pillow
playwright
aiofiles
pytest

# Dependencia para parseo de IP en argparse
ipaddress